ACTIN-2464 Container and README for ACTIN
Add a `Dockerfile` to make running the `actin_curator` easy. Potentially
in future this same file could be used to run the other entry points but
for now I focused on the the unmet need for ACTIN.

Also add a `requirements.txt` for `actin_curator` which describes all
required dependencies and a wrapper script, which should work for ACTIN
members to run the container locally.

The wrapper script tries to create a local directory tree which will
contain the input and output files in well-known locations. It also
endeavours to locate the configuration directory for the `gcloud`
command on the user's host.

It bind-mounts both directories into the container on startup and
executes the Python code within.